Do blacklisting in more appropriate spot
